pyproject.toml
[project]
name = "apify-scrape-social"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=1.12.0",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "supabase>=2.17.0",
    "uvicorn[standard]>=0.35.0",
]
